Decouple Artifact building from the custom built Resolver.
This is a part of my effort to replace the Resolver, with Aether.
Currently, artifacts are built using Resolver.getArtifact().
In addition, the Resolver defines a class of exceptions that is
associated with creating a new DefaultArtifact.
This functionality can and should be removed from the Resolver.

After this change, the resolver is only called from within
(1) itself, (2) its tests and (3) the defaultmodelresolver.