Resolve versions using aether rather than adhoc solution.
This is a part of an ongoing effort to isolate and then replace
the existing Resolver with Aether.

Version resolution is the only remaining part of the Resolver
that is used by any other class, DefaultModelResolver. With this
removed the Resolver is currently only used by itself, its tests and
GenerateWorkspace.

The current implementation of resolving version given a version spec
is adhoc and idiosyncratic. This CL isolates this functionality, and
provides a canonical/correct version resolution mechanism.

Correct version resolution has two components. (1) Pins both hard and
soft, e.g. "[3.0]" and "3.0", should select the pinned version, e.g.
"3.0", and (2) version ranges should select the highest available
version.

This functionality is easily accomplished through aether.
It involves the construction of a version range request,
followed by a call to aether.

Tests are implemented for when there is invalid behavior, as well
as for pinned version ranges, i.e. "3.2". However, due to limitations
of Mockito and Aether (mockito annot mock final classes), tests
are not included for correct behavior of version ranges.