Resolve versions using aether, and uncouple existing Resolver from code base. (#45)
* Decouple Artifact building from the custom built Resolver.

This is a part of my effort to replace the Resolver, with Aether.
Currently, artifacts are built using Resolver.getArtifact().
In addition, the Resolver defines a class of exceptions that is
associated with creating a new DefaultArtifact.
This functionality can and should be removed from the Resolver.

After this change, the resolver is only called from within
(1) itself, (2) its tests and (3) the defaultmodelresolver.

* Resolve versions using aether rather than adhoc solution.

This is a part of an ongoing effort to isolate and then replace
the existing Resolver with Aether.

Version resolution is the only remaining part of the Resolver
that is used by any other class, DefaultModelResolver. With this
removed the Resolver is currently only used by itself, its tests and
GenerateWorkspace.

The current implementation of resolving version given a version spec
is adhoc and idiosyncratic. This CL isolates this functionality, and
provides a canonical/correct version resolution mechanism.

Correct version resolution has two components. (1) Pins both hard and
soft, e.g. "[3.0]" and "3.0", should select the pinned version, e.g.
"3.0", and (2) version ranges should select the highest available
version.

This functionality is easily accomplished through aether.
It involves the construction of a version range request,
followed by a call to aether.

Tests are implemented for when there is invalid behavior, as well
as for pinned version ranges, i.e. "3.2". However, due to limitations
of Mockito and Aether (mockito annot mock final classes), tests
are not included for correct behavior of version ranges.

* Resolve soft pinned version to nearest version if it does not exist.

Currently, even if a soft pinned version, e.g. "3.0", does not
exist aether returns that soft pinned version. Instead, it should
provide another valid version. This change solves this issue by
transforming the soft pin into a version range, and making the
aether request using that version range.

For example, instead of requesting for "3.0", we transform it to "[3.0,)"
providing us all versions between 3.0 and the latest version.
If the version exists, we simply take it. If it doesn't, then
we select the next closest valid version.